chore: move class based UserInputManager controllers into 'service' directory
